feat(usePagination): implement usePagination and apply to index
