+++ gatsby_site/src/pages/index.js
@@ -1,11 +1,11 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import styled from 'styled-components';
+import SearchButton from '../components/BaseComponents/SearchButton';
 import EventsContainer from '../components/EventsContainer';
 import Layout from '../components/layout';
 import useLocalStorage from '../hooks/useLocalStorage';
+import usePagination from '../hooks/usePagination';
 import useStaticResources from '../hooks/useStaticResources';
-import { Events as evUtil } from '../utils';
-import SearchButton from '../components/BaseComponents/SearchButton';
 
 const HeaderWrapper = styled.div`
   display: flex;
@@ -18,37 +18,8 @@ export default () => {
   useLocalStorage();
   useStaticResources();
 
-  // local state
-  const [isLoading, setIsLoading] = useState(false);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [fetchedData, setFetchedData] = useState({});
-  const [currentPageEvents, setCurrentPageEvents] = useState([]);
-
-  useEffect(() => {
-    setIsLoading(true);
-    fetch(`/api/pagination?page=${currentPage}`)
-      .then(res => res.json())
-      .then((resJson) => {
-        setFetchedData(resJson);
-        setIsLoading(false);
-      });
-  }, [currentPage]);
-
-  useEffect(() => {
-    if (fetchedData.data) {
-      // preprocess events before sending to events list
-      const preprocessedEvents = evUtil.getPreprocessedEvents(
-        fetchedData.data
-          // TODO: this filter step is supposed to be done on backend
-          .filter(x => x.title)
-          .map(x => ({ node: x })),
-        true,
-      );
-
-      // set to local state
-      setCurrentPageEvents(preprocessedEvents);
-    }
-  }, [fetchedData.data]);
+  // use a custom hook to get data
+  const [currentPageEvents, totalPage, setCurrentPage, isLoading] = usePagination();
 
   return (
     <Layout>
@@ -60,7 +31,7 @@ export default () => {
         isLoading={isLoading}
         allEvents={currentPageEvents}
         handlePagination={pageNo => setCurrentPage(pageNo)}
-        noOfPages={fetchedData.total}
+        noOfPages={totalPage}
         currentPageEvents={currentPageEvents}
       />
     </Layout>
